refactor: code tests configurations refactor and init integrations test ♻️
