+++ infrasctructure/src/test/java/com/curso/admin/catalogo/CleanUpExtension.java
@@ -1,0 +1,26 @@
+package com.curso.admin.catalogo;
+
+import org.junit.jupiter.api.extension.BeforeEachCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import java.util.Collection;
+
+public class CleanUpExtension implements BeforeEachCallback {
+
+    @Override
+    public void beforeEach(ExtensionContext context) throws Exception {
+        final var repositories = SpringExtension
+                .getApplicationContext(context)
+                .getBeansOfType(CrudRepository.class)
+                .values();
+
+        cleanUp(repositories);
+    }
+
+    private void cleanUp (final Collection<CrudRepository> repositories) {
+        repositories.forEach(CrudRepository::deleteAll);
+    }
+}
+
