+++ infrasctructure/src/test/java/com/curso/admin/catalogo/IntegrationTest.java
@@ -1,0 +1,18 @@
+package com.curso.admin.catalogo;
+
+import com.curso.admin.catalogo.infrasctructure.configuration.WebServerConfig;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import org.springframework.test.context.ActiveProfiles;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Inherited
+@ActiveProfiles("test")
+@SpringBootTest(classes = WebServerConfig.class)
+@ExtendWith(CleanUpExtension.class)
+public @interface IntegrationTest {
+}
